chore: update story title
