+++ docs/typography/examples/TextStylesPage.stories.tsx
@@ -1,11 +1,13 @@
-import { ComponentMeta, ComponentStory } from '@storybook/react';
+import { ComponentStory } from '@storybook/react';
 
 import TextStylesPage from './TextStylesPage';
 export default {
-  title: 'Components/Typography',
+  title: 'Typography/Examples',
   component: TextStylesPage,
 };
 
 const Template: ComponentStory<typeof TextStylesPage> = () => <TextStylesPage />;
 
 export const TextStyles = Template.bind({});
+
+TextStyles.args = {};
