+++ tsconfig.json
@@ -19,7 +19,11 @@
     "jsx": "react-jsx",
     "types": ["vite/client"]
   },
-  "include": ["src"],
+  "include": [
+    "src",
+    "docs/components/Typography/TextStylesPage.tsx",
+    "docs/components/Typography/TextStylesPage.stories.tsx"
+  ],
   "exclude": ["**/*.stories.*"],
   "references": [{ "path": "./tsconfig.node.json" }]
 }
